requirements.txt
anyio==4.12.1
certifi==2026.2.25
charset-normalizer==3.4.6
configparser==7.2.0
dnspython==2.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pymongo==4.16.0
python-telegram-bot==22.5
requests==2.32.5
urllib3==2.6.2
